data, yaml: store data step info on Thing constructors
